api/requirements.txt
# 1id.com Enrollment API dependencies
# Port 8180 - see /ports.md for registry
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
mysql-connector-python>=8.0.0
cryptography>=36.0.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
